mcux: drivers: imx: fix spi timing delay overflow
define realDelay by uint32_t may be overflow.
overflow will cause delay cal error.

+++ mcux/drivers/imx/fsl_lpspi.c
@@ -577,7 +577,7 @@ uint32_t LPSPI_MasterSetDelayTimes(LPSPI_Type *base,
                                    lpspi_delay_type_t whichDelay,
                                    uint32_t srcClock_Hz)
 {
-    uint32_t realDelay, bestDelay;
+    uint64_t realDelay, bestDelay;
     uint32_t scaler, bestScaler;
     uint32_t diff, min_diff;
     uint64_t initialDelayNanoSec;
